[refactor] :recycle: rename states.rs to game.rs and move inside state folder
