+++ anchor/programs/lottery/src/state/game.rs
