+++ anchor/programs/lottery/src/state/mod.rs
@@ -1,0 +1,1 @@
+pub mod game;
